# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "향수 정보 서비스 A.fume Server Api 문서"
  version: 1.0.0
  title: "Swagger A.fume Server"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "heesung6701@naver.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "{SERVER_URL}:{SERVER_PORT}"
basePath: "/A.fume/api/0.0.1"
...
schemes:
#- "https"
- "http"
# security:
# - APIKeyHeader: []
paths:
  /perfume:
    post:
      tags:
      - "perfume"
      summary: "향수 정보 추가"
      description: "향수 정보를 추가한다."
      operationId: "createPerfume"
      security:
        - userToken: []
      x-security-scopes:
        - admin
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          allOf:
          - $ref: "#/definitions/PerfumeInfo"
          - $ref: "#/definitions/PerfumeDetailInfo"
      responses:
        "200":
          description: "성공"
          schema:
            type: "object"
            properties:
              message:
                type: string
                example: 향수 추가 성공
              data:
                type: integer
                example: 1
                description: 향수 perfumeIdx
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
        "405":
          description: "Invalid input"
      deprecated: true
      x-swagger-router-controller: "Perfume"
  /perfume/search:
    post:
      tags:
      - "perfume"
      summary: "향수 검색"
      description: "검색 조건에 해당하는 향수를 반환한다."
      operationId: "searchPerfume"
      security:
        - userToken: []
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "filter"
        description: "검색 필터"
        required: false
        schema:
          $ref: "#/definitions/Filter"
      responses:
        "200":
          description: 성공
          schema:
            type: object
            properties:
              message:
                type: string
                example: 향수 검색 성공
              data:
                type: array
                items:
                  $ref: "#/definitions/Perfume"
        "400":
          description: "Invalid tag value"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      x-swagger-router-controller: "Perfume"
  /perfume/recentSearch:
    get:
      tags:
      - perfume
      summary: 최근 검색한 향수 조회
      description: 최근에 검색한 향수를 조회한다.
      operationId: recentSearch
      security:
        - userToken: []
      x-security-scopes:
        - user
      produces:
      - application/json
      responses:
        200:
          description: 성공
          schema:
            type: object
            properties:
              message:
                type: string
                example: 향수 검색 성공
              data:
                type: array
                items:
                  $ref: "#/definitions/Perfume"
        400:
          description: Invalid tag value
        401:
          description: API key is missing or invalid
          headers:
            WWW_Authenticate:
              type: string
      x-swagger-router-controller: Perfume
  /perfume/{perfumeIdx}:
    get:
      tags:
      - "perfume"
      summary: "향수 세부 정보 조회"
      description: "향수 세부 정보를 반환한다."
      security:
        - userToken: []
      operationId: "getPerfumeById"
      produces:
      - "application/json"
      parameters:
      - name: "perfumeIdx"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
          schema:
            type: object
            properties:
              message:
                type: string
                example: 향수 세부 조회 성공
              data:
                $ref: "#/definitions/PerfumeDetail"
        "401":
          description: "Not authenticated"
        "403":
          description: "Access token does not have the required scope"
        "404":
          description: "Perfume not found"
      x-swagger-router-controller: "Perfume"
    put:
      tags:
      - "perfume"
      summary: "향수 정보 업데이트"
      operationId: "updatePerfume"
      security:
        - userToken: []
      x-security-scopes:
        - admin
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "perfumeIdx"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/PerfumeInfo"
      responses:
        "200":
          description: "성공"
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
        "405":
          description: "Invalid input"
      deprecated: true
      x-swagger-router-controller: "Perfume"
    delete:
      tags:
      - "perfume"
      summary: "Deletes a perfume"
      operationId: "deletePerfume"
      security:
        - userToken: []
      x-security-scopes:
        - admin
      produces:
      - "application/json"
      parameters:
      - name: "perfumeIdx"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
        "404":
          description: "Perfume not found"
      deprecated: true
      x-swagger-router-controller: "Perfume"
  /perfume/{perfumeIdx}/like:
    put:
      tags:
      - "perfume"
      summary: "향수 좋아요"
      description: "향수 좋아요 상태를 반환한다."
      security:
        - userToken: []
      x-security-scopes:
        - user
      operationId: "likePerfume"
      produces:
      - "application/json"
      parameters:
      - name: "perfumeIdx"
        in: "path"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
          schema:
            type: object
            properties:
              message:
                type: string
                example: 향수 세부 조회 성공
              data:
                type: boolean
                example: true
                description: 요청 이후 좋아요 상태
        "401":
          description: "Not authenticated"
        "403":
          description: "Access token does not have the required scope"
        "404":
          description: "Perfume not found"
      x-swagger-router-controller: "Perfume"
...
  /review/{reviewIdx}/reply:
    post:
      tags:
      - "reply"
      summary: "댓글 추가"
      description : "시향노트에 댓글 추가하기"
      operationId: "postReply"
      security:
        - userToken: []
      x-security-scopes:
        - user
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "reviewIdx"
        in: "path"
        description: "시향노트 Idx"
        required: true
        type: "integer"
        format: "int64"
      - name: "body"
        in: "body"
        description: "댓글 정보"
        required: true
        schema:
          $ref: "#/definitions/ReplyInfo"
      responses:
        "200":
          description: "성공"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "Reply"
    get:
      tags:
      - "reply"
      summary: "전체 댓글 반환"
      description: "특정 시향노트에 달린 전체 댓글 가져오기"
      operationId: "getReplyOfReview"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "reviewIdx"
        in: "path"
        description: "시향노트 Idx"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              allOf:
              - $ref: "#/definitions/Reply"
              - type: "object"
                properties:
                  replyIdx:
                    type: "integer"
                    format: "int64"
                example:
                  replyIdx: 1
        "400":
          description: "Invalid username/password supplied"
      x-swagger-router-controller: "Reply"
  /reply/{replyIdx}:
    get:
      tags:
      - "reply"
      summary: "댓글 반환"
      description: "시향노트에 달린 특정 댓글 가져오기"
      operationId: "getReplyByIdx"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "replyIdx"
        in: "path"
        description: "댓글 Idx"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
          schema:
            $ref: "#/definitions/Reply"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
        "404":
          description: "Reply not found"
      x-swagger-router-controller: "Reply"
    put:
      tags:
      - "reply"
      summary: "댓글 수정"
      description: "댓글 수정하기"
      operationId: "updateReply"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "replyIdx"
        in: "path"
        description: "댓글 Idx"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/ReplyInfo"
      responses:
        "200":
          description: "성공"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
        "405":
          description: "Invalid input"
      x-swagger-router-controller: "Reply"
    delete:
      tags:
      - "reply"
      summary: "댓글 삭제"
      description: "댓글 삭제하기"
      operationId: "deleteReply"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "replyIdx"
        in: "path"
        description: "댓글 Idx"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "성공"
        "400":
          description: "Invalid ID supplied"
        "401":
          description: "API key is missing or invalid"
          headers:
            WWW_Authenticate:
              type: "string"
      x-swagger-router-controller: "Reply"
...
securityDefinitions:
  userToken:
      type: apiKey
      name: x-access-token
      in: header
      description: /Bearer {token}/ 를 입력해주시면 됩니다.
  APIKeyHeader:
    type: apiKey
    name: X-API-Key
    in: header
definitions:
...
  PerfumeInfo:
    type: "object"
    properties:
      name:
        type: string
      englishName:
        type: string
      mainSeriesIdx:
        type: integer
      brandIdx:
        type: integer
      imageUrl:
        type: string
      releaseDate:
        type: string
    example:
      name: 154 코롱
      englishName: 154 Cologne
      mainSeriesIdx: 1
      brandIdx: 1
      imageUrl: https://contents.lotteon.com/itemimage/_v065423/LE/12/04/59/50/19/_1/22/48/08/13/9/LE1204595019_1224808139_1.jpg/dims/resizef/554X554
      releaseDate: "2020-11-20"
  Perfume:
    properties:
      perfumeIdx:
        type: integer
      likeCnt:
        type: integer
      isLiked:
        type: boolean
      brandName:
        type: string
      mainSeriesName:
        type: string
    allOf:
    - $ref: "#/definitions/PerfumeInfo"
    - {}
    example:
      perfumeIdx: 1
      likeCnt: 4
      isLiked: true
      brandName: "딥디크 파리스"
      mainSeriesName: "플로럴"
  PerfumeDetailInfo:
    properties:
      story:
        type: string
      abundanceRate:
        type: string
        enum:
        - 코롱
        - 오 드 뜨왈렛
        - 오 드 퍼퓸
        - 퍼퓸
      imageUrl:
        type: string
      volumeAndPrice:
        type: array
        items:
          type: "object"
          properties:
            volume:
              type: integer
            price:
              type: integer
    example:
      story: "조 말론 런던 1호점이 위치한 런던의 거리 번호입니다. 광범위한 후각적 탐구를 요하는 이 향수는 만다린, 그레이프 프루트\
        , 바질, 너트맥, 베티버와 같은 브랜드를 대표하는 성분들을 모두 함유하고 있습니다. 다양한 느낌을 연출하는 향입니다."
      abundanceRate: 코롱
      volumeAndPrice:
      - volume: 30
        price: 90000
      - volume: 30
        price: 90000
      - volume: 30
        price: 90000
  PerfumeDetail:
    properties:
      score:
        type: number
        description: 점수 평균 값
        minimum: 0
        maximum: 5
      longevity:
        type: object
        description: 지속감
        properties:
          veryWeak:
            type: number
            minimum: 0
            maximum: 1
            description: 매우 약함
          weak:
            type: number
            minimum: 0
            maximum: 1
            description: 약함
          medium:
            type: number
            minimum: 0
            maximum: 1
            description: 보통
          strong:
            type: number
            minimum: 0
            maximum: 1
            description: 강함
          veryStrong:
            type: number
            minimum: 0
            maximum: 1
            description: 매우 강함
      sillage:
        type: object
        description: 잔향감
        properties:
          light:
            type: number
            minimum: 0
            maximum: 1
            description: 약함
          normal:
            type: number
            minimum: 0
            maximum: 1
            description: 보통
          heavy:
            type: number
            minimum: 0
            maximum: 1
            description: 강함
      seasonal:
        type: object
        properties:
          spring:
            type: number
            minimum: 0
            maximum: 1
            description: 봄
          summer:
            type: number
            minimum: 0
            maximum: 1
            description: 여름
          fall:
            type: number
            minimum: 0
            maximum: 1
            description: 가을
          winter:
            type: number
            minimum: 0
            maximum: 1
            description: 겨울
      keywords:
        type: "array"
        items:
          type: "string"
          example: ""
      noteType:
        type: "integer"
        description: "0은 일반 노트, 1은 single 노트"
      ingredients:
        $ref: "#/definitions/PerfumeDetailInfo_ingredients"
    allOf:
    - $ref: "#/definitions/PerfumeInfo"
    - $ref: "#/definitions/PerfumeDetailInfo"
    - type: "object"
      properties:
        perfumeIdx:
          type: "integer"
      example:
        perfumeIdx: 1
    example:
      score: 5.4
      abundanceRate: "오 드 퍼퓸"
      seasonal:
        spring: 0
        summer: 0.3
        fall: 0.3
        winter: 0.4
      sillage:
        light: 0.4
        medium: 0.2
        heavy: 0.4
      longevity:
        veryWeak: 0.1
        weak: 0.1
        normal: 0.2
        strong: 0.2
        veryStrong: 0.1
      noteType: 0
      ingredients:
        top:
        - ingredientIdx: 1
          name: "재료1"
          description: "향긋"
        - ingredientIdx: 5
          name: "재료5"
          description: "향긋"
        middle:
        - ingredientIdx: 2
          name: "재료2"
          description: "향긋"
        - ingredientIdx: 6
          name: "재료6"
          description: "향긋"
        base:
        - ingredientIdx: 3
          name: "재료3"
          description: "향긋"
        - ingredientIdx: 4
          name: "재료4"
          description: "향긋"
        - ingredientIdx: 7
          name: "재료7"
          description: "향긋"
...
  IngredientInfo:
    type: "object"
    properties:
      name:
        type: "string"
      englishName:
        type: "string"
      description:
        type: "string"
    example:
      name: "씨쏠트"
      englishName: "Sea Salt"
      description: "바다 소금"
...
  ReplyInfo:
    type: "object"
    properties:
      userIdx:
        type: "integer"
        format: "int64"
      content:
        type: "string"
    example:
      userIdx: 2
      content: "저한테도 인생 향수에요~"
  Reply:
    type: "object"
    properties:
      nickname:
        type: "string"
        example: 
    allOf:
    - $ref: "#/definitions/ReplyInfo"
    - {}
    example:
      nickname: "향린이S2"
...
  Filter:
    type: "object"
    properties:
      series:
        type: "array"
        items:
          type: "string"
      brands:
        type: "array"
        items:
          type: "string"
      keywords:
        type: "array"
        items:
          type: "string"
      sortBy:
        type: "string"
    example:
      series:
      - 플로럴
      - 오리엔탈
      - 시트러스
      brands:
      - 조 말론 런던
      - 르 라보
      - 딥디크 파리스
      keywords:
      - 구현전
      sortBy: 인기순
...
  ApiResponse:
    type: "object"
    properties:
      message:
        type: "string"
        example: 메세지
      data:
        type: "string"
        example: 0
  PerfumeDetailInfo_ingredients:
    type: "object"
    properties:
      top:
        type: "array"
        items:
          $ref: "#/definitions/IngredientInfo"
      middle:
        type: "array"
        items:
          $ref: "#/definitions/IngredientInfo"
      base:
        type: "array"
        items:
          $ref: "#/definitions/IngredientInfo"
      single:
        type: "array"
        items:
          $ref: "#/definitions/IngredientInfo"
...